Clean up view registration
Registration is now a function of the render module, since this is
really a whitelist for who gets to be rendered.

A decorated view function is exported from render module that will
define and register a view in one go.
