+++ typescript/packages/common-ui/src/components/div.ts
@@ -1,5 +1,4 @@
-import { view } from '../hyperscript/view.js';
-import { register as registerView } from '../hyperscript/known-tags.js';
+import { view } from '../hyperscript/render.js';
 
 export const div = view('div', {
   type: 'object',
@@ -7,5 +6,3 @@ export const div = view('div', {
     id: { type: 'string' },
   }
 });
-
-registerView(div);
