+++ typescript/packages/common-ui/src/components/hstack.ts
@@ -1,11 +1,8 @@
 import { LitElement, html, css } from 'lit-element';
 import { customElement } from 'lit-element/decorators.js';
-import { view } from '../hyperscript/view.js';
-import { register as registerView } from '../hyperscript/known-tags.js';
+import { view } from '../hyperscript/render.js';
 
 export const hstack = view('co-hstack', {});
-
-registerView(hstack);
 
 @customElement('co-hstack')
 export class HStackElement extends LitElement {
