+++ typescript/packages/common-ui/src/components/navpanel.ts
@@ -1,11 +1,9 @@
 import { LitElement, html, css } from 'lit-element';
 import { customElement, property } from 'lit-element/decorators.js';
-import { VNode, view } from '../hyperscript/view.js';
-import { render, RenderContext } from '../hyperscript/render.js';
-import { register as registerView } from '../hyperscript/known-tags.js';
+import { VNode } from '../hyperscript/view.js';
+import { render, RenderContext, view } from '../hyperscript/render.js';
 
 export const navpanel = view('co-navpanel', {});
-registerView(navpanel);
 
 @customElement('co-navpanel')
 export class NavPanelElement extends LitElement {
