+++ typescript/packages/common-ui/src/hyperscript/render.ts
@@ -1,7 +1,36 @@
 import { Cancel, combineCancels } from '@commontools/common-frp';
 import { Signal, effect } from '@commontools/common-frp/signal';
-import { isSignalBinding, VNode } from './view.js';
-import { getViewByTag } from './known-tags.js';
+import {
+  isSignalBinding,
+  VNode,
+  AnyJSONSchema,
+  View,
+  view as createView
+} from './view.js';
+
+const registry = () => {
+  const viewByTag = new Map<string, View>();
+
+  const getViewByTag = (tag: string) => viewByTag.get(tag);
+
+  const register = (view: View) => {
+    viewByTag.set(view.tag, view);
+  }
+
+  return {getViewByTag, register};
+}
+
+export const {getViewByTag, register} = registry();
+
+/** Define and register a view factory function */
+export const view = (
+  tagName: string,
+  propsSchema: AnyJSONSchema = {}
+): View => {
+  const factory = createView(tagName, propsSchema);
+  register(factory);
+  return factory;
+}
 
 export type RenderContext = Record<string, Signal<any>>
